better Install.py script for KIM from Ryan Elliott

+++ lib/kim/Install.py
@@ -1,25 +1,44 @@
 #!/usr/bin/env python
 
-# install.py tool to setup the kim-api library
+# install.pa tool to setup the kim-api library
 # used to automate the steps described in the README file in this dir
 
 import sys,os,re,urllib,commands
 
 help = """
-Syntax: install.py -v version -c kim-dir -b kim-model-name -a kim-name
-        specify one or more options, order does not matter
-        -v = version of kim-api to download and work with
-             default = kim-api-v1.8.2 (current as of June 2017)
-        -c = create Makefile.KIM_DIR within lammps lib/kim to configure lammps
-             for use with the kim-api library installed at "kim-dir" (absolute
-             path).  default = this dir
-        -b = build kim-api and kim model where kim-model-name can be a specific
-             openkim.org model name (such as
-             "EAM_Dynamo_Ackland_W__MO_141627196590_002") or the keyword
-             "OpenKIM" to install all compatible models from the openkim.org
-             site.
-        -a = add kim-name openkim.org item (model driver or model) to existing
-             kim-api instalation.
+Syntax from src dir: make lib-kim args="-v version -b kim-install-dir kim-name -a kim-name"
+Syntax from lib dir: python Install.py -v version -b kim-install-dir kim-name -a kim-name
+
+specify one or more options, order does not matter
+
+  -v = version of KIM API library to use
+       default = kim-api-v1.8.2 (current as of June 2017)
+  -b = download and build KIM API library with KIM models
+       kim-dir = where to install/build the KIM API library
+         use "." to install in lib/kim
+       kim-name = none to install only the example KIM models
+       kim-name = KIM model name (see example below) + examples
+       kim-name = OpenKIM to install all models
+         from the openkim.org site (this can take 30 minutes or more)
+  -a = add single KIM model or model driver with kim-name
+       to existing KIM API lib (see example below)
+
+Examples:
+
+make lib-kim args="-b . none"   # install KIM API lib with only example models
+make lib-kim args="-b . Glue_Ercolessi_Adams_Al__MO_324507536345_001"  # ditto plus one model
+make lib-kim args="-b . OpenKIM"   # install KIM API lib with all models
+make lib-kim args="-a EAM_Dynamo_Ackland_W__MO_141627196590_002"   # add one model or model driver
+
+See the list of KIM model drivers here:
+https://openkim.org/kim-items/model-drivers/alphabetical
+
+See the list of all KIM models here:
+https://openkim.org/kim-items/models/by-model-drivers
+
+See the list of example KIM models included by default here:
+https://openkim.org/kim-api
+in the "What is in the KIM API source package?" section
 """
 
 def error():
@@ -28,32 +47,28 @@ def error():
 
 # parse args
 
-args = sys.argv
+args = sys.argv[1:]
+nargs = len(args)
+if nargs == 0: error()
 
 thisdir = os.environ['PWD']
-dir = thisdir
 version = "kim-api-v1.8.2"
 
-dirflag = 0
 buildflag = 0
 addflag = 0
 
-iarg = 1
+iarg = 0
 while iarg < len(args):
   if args[iarg] == "-v":
     if iarg+2 > len(args): error()
     version = args[iarg+1]
     iarg += 2
-  elif args[iarg] == "-c":
-    dirflag = 1
-    if iarg+2 > len(args): error()
-    dir = args[iarg+1]
-    iarg += 2
   elif args[iarg] == "-b":
     buildflag = 1
-    if iarg+2 > len(args): error()
-    modelname = args[iarg+1]
-    iarg += 2
+    if iarg+3 > len(args): error()
+    dir = args[iarg+1]
+    modelname = args[iarg+2]
+    iarg += 3
   elif args[iarg] == "-a":
     addflag = 1
     if iarg+2 > len(args): error()
@@ -62,33 +77,59 @@ while iarg < len(args):
   else: error()
 
 thisdir = os.path.abspath(thisdir)
-dir = os.path.abspath(dir)
 url = "https://s3.openkim.org/kim-api/%s.tgz" % version
 
-# download and unpack tarball
-
-
-if not os.path.isfile("%s/Makefile.KIM_DIR" % thisdir):
-  open("%s/Makefile.KIM_DIR" % thisdir, 'w').write("KIM_INSTALL_DIR=%s" % dir)
-  open("%s/Makefile.KIM_Config" % thisdir, 'w').write("include %s/lib/kim-api/Makefile.KIM_Config" % dir)
-  print "Created %s/Makefile.KIM_DIR : using %s" % (thisdir,dir)
-else:
-  if dirflag == 1:
-    open("%s/Makefile.KIM_DIR" % thisdir, 'w').write("KIM_INSTALL_DIR=%s" % dir)
-    open("%s/Makefile.KIM_Config" % thisdir, 'w').write("include %s/lib/kim-api/Makefile.KIM_Config" % dir)
-    print "Updated %s/Makefile.KIM_DIR : using %s" % (thisdir,dir)
-
+# download KIM tarball, unpack, build KIM
+# either in lib/kim or user-requested location
 
 if buildflag == 1:
-  # download kim-api
+
+  # set install directory
+  dir = os.path.abspath(dir) + "/installed-" + version
+
+  # check to see if an installed kim-api already exists
+
+  if os.path.isdir(dir):
+     print "kim-api is already installed at %s" % dir
+     print "Must remove this directory in order to resintall at this location"
+     sys.exit()
+
+  # configure LAMMPS to use kim-api to be installed
+
+  mkfle = open("%s/Makefile.KIM_DIR" % thisdir, 'w')
+  mkfle.write("KIM_INSTALL_DIR=%s\n" % dir)
+  mkfle.write("\n")
+  mkfle.write(".DUMMY: print_dir\n")
+  mkfle.write("\n")
+  mkfle.write("print_dir:\n")
+  mkfle.write("	@printf $(KIM_INSTALL_DIR)\n")
+  mkfle.close()
+  open("%s/Makefile.KIM_Config" % thisdir, 'w'). \
+    write("include %s/lib/kim-api/Makefile.KIM_Config" % dir)
+  print "Created %s/Makefile.KIM_DIR : using %s" % (thisdir,dir)
+
+  # download entire kim-api tarball
+  # try first via urllib
+  # if fails (probably due to no SSL support), use wget
+
   print "Downloading kim-api tarball ..."
-  urllib.urlretrieve(url,"%s/%s.tgz" % (thisdir,version))
+
+  try: urllib.urlretrieve(url,"%s/%s.tgz" % (thisdir,version))
+  except:
+    cmd = "wget %s %s/%s.tgz" % (url,thisdir,version)
+    txt = commands.getstatusoutput(cmd)
+    print txt[1]
+    if not os.path.isfile("%s/%s.tgz" % (thisdir,version)):
+      print "Both urllib.urlretrieve() and wget command failed to download"
+      sys.exit()
+
   print "Unpacking kim-api tarball ..."
   cmd = "cd %s; rm -rf %s; tar zxvf %s.tgz" % (thisdir,version,version)
   txt = commands.getstatusoutput(cmd)
   if txt[0] != 0: error()
 
   # configure kim-api
+
   print "Configuring kim-api ..."
   cmd = "cd %s/%s; ./configure --prefix='%s'" % (thisdir,version,dir)
   txt = commands.getstatusoutput(cmd)
@@ -96,12 +137,19 @@ if buildflag == 1:
   if txt[0] != 0: error()
 
   # build kim-api
+
   print "Configuring model : %s" % modelname
-  cmd = "cd %s/%s; make add-%s" % (thisdir,version,modelname)
+  if modelname == "none":
+    cmd = "cd %s/%s; make add-examples" % (thisdir,version)
+  else:
+    if modelname == "OpenKIM":
+      print "configuring all OpenKIM models, this will take a while ..."
+    cmd = "cd %s/%s; make add-examples; make add-%s" % \
+        (thisdir,version,modelname)
   txt = commands.getstatusoutput(cmd)
   print txt[1]
   if txt[0] != 0: error()
-  #
+
   print "Building kim-api ..."
   cmd = "cd %s/%s; make" % (thisdir,version)
   txt = commands.getstatusoutput(cmd)
@@ -109,42 +157,101 @@ if buildflag == 1:
   if txt[0] != 0: error()
 
   # install kim-api
+
   print "Installing kim-api ..."
   cmd = "cd %s/%s; make install" % (thisdir,version)
   txt = commands.getstatusoutput(cmd)
   print txt[1]
   if txt[0] != 0: error()
-  #
+
   cmd = "cd %s/%s; make install-set-default-to-v1" %(thisdir,version)
   txt = commands.getstatusoutput(cmd)
   print txt[1]
   if txt[0] != 0: error()
 
   # remove source files
+
   print "Removing kim-api source and build files ..."
   cmd = "cd %s; rm -rf %s; rm -rf %s.tgz" % (thisdir,version,version)
   txt = commands.getstatusoutput(cmd)
   print txt[1]
   if txt[0] != 0: error()
 
+# add a single model (and possibly its driver) to existing KIM installation
+
 if addflag == 1:
-  # download model
-  url = "https://openkim.org/download/%s.tgz" % addmodelname
+
+  # get location of installed kim-api
+
+  if not os.path.isfile("%s/Makefile.KIM_DIR" % thisdir):
+    print "kim-api is not installed"
+    error()
+  else:
+    cmd = "cd %s; make -f Makefile.KIM_DIR print_dir" % thisdir
+    dir = commands.getstatusoutput(cmd)[1]
+
+  # download single model
+  # try first via urllib
+  # if fails (probably due to no SSL support), use wget
+
   print "Downloading item tarball ..."
-  urllib.urlretrieve(url,"%s/%s.tgz" % (thisdir,addmodelname))
+
+  url = "https://openkim.org/download/%s.tgz" % addmodelname
+
+  try: urllib.urlretrieve(url,"%s/%s.tgz" % (thisdir,addmodelname))
+  except:
+    cmd = "wget %s %s/%s.tgz" % (url,thisdir,addmodelname)
+    txt = commands.getstatusoutput(cmd)
+    print txt[1]
+    if not os.path.isfile("%s/%s.tgz" % (thisdir,addmodelname)):
+      print "Both urllib.urlretrieve() and wget command failed to download"
+      sys.exit()
+
   print "Unpacking item tarball ..."
   cmd = "cd %s; tar zxvf %s.tgz" % (thisdir,addmodelname)
   txt = commands.getstatusoutput(cmd)
   if txt[0] != 0: error()
-  #
+
   print "Building item ..."
   cmd = "cd %s/%s; make; make install" %(thisdir,addmodelname)
   txt = commands.getstatusoutput(cmd)
-  print txt[1]
-  if txt[0] != 0: error()
-  #
-  print "Removing kim item source and build files ..."
-  cmd = "cd %s; rm -rf %s; rm -rf %s.tgz" %(thisdir,addmodelname,addmodelname)
-  txt = commands.getstatusoutput(cmd)
-  print txt[1]
-  if txt[0] != 0: error()
+  firstRunOutput = txt[1]
+  if txt[0] != 0:
+    # Error: but first, check to see if it needs a driver
+
+    cmd = "cd %s/%s; make kim-item-type" % (thisdir,addmodelname)
+    txt = commands.getstatusoutput(cmd)
+    if txt[1] == "ParameterizedModel":
+
+      # Get and install driver
+
+      cmd = "cd %s/%s; make model-driver-name" % (thisdir,addmodelname)
+      txt = commands.getstatusoutput(cmd)
+      adddrivername = txt[1]
+      print "First Installing model driver: %s" % adddrivername
+      cmd = "cd %s; python Install.py -a %s" % (thisdir,adddrivername)
+      txt = commands.getstatusoutput(cmd)
+      if txt[0] != 0:
+         print firstRunOutput
+         print txt[1]
+         error()
+      else:
+        print txt[1]
+      cmd = "cd %s; python Install.py -a %s" % (thisdir,addmodelname)
+      txt = commands.getstatusoutput(cmd)
+      print txt[1]
+      if txt[0] != 0:
+        error()
+    else:
+      print firstRunOutput
+      error()
+  else:
+
+    # success
+
+    print firstRunOutput
+    print "Removing kim item source and build files ..."
+    cmd = "cd %s; rm -rf %s; rm -rf %s.tgz" %(thisdir,addmodelname,addmodelname)
+    txt = commands.getstatusoutput(cmd)
+    print txt[1]
+    if txt[0] != 0: error()
